some better condition checking

+++ rdfdot-core/src/main/native/build-graphviz.sh
@@ -17,11 +17,22 @@ mkdir -p $WORKDIR
 mkdir -p $DLDIR
 
 GRAPHVIZ_HREF="http://www.graphviz.org/pub/graphviz/stable/SOURCES/graphviz-2.36.0.tar.gz"
+GRAPHVIZ_MD5=1f41664dba0c93109ac8b71216bf2b57
+
 LIBGD_HREF="https://bitbucket.org/libgd/gd-libgd/downloads/libgd-2.1.0.tar.gz"
+LIBGD_MD5=20a8eeae51cef7d7850679b1e53ce2b5
+
 LIBPNG_HREF="http://freefr.dl.sourceforge.net/project/libpng/libpng12/1.2.51/libpng-1.2.51.tar.gz"
+LIBPNG_MD5=e358f9a265f2063b36f10dc454ee0e17
+
 LIBJPEG_HREF="http://www.ijg.org/files/jpegsrc.v9a.tar.gz"
+LIBJPEG_MD5=3353992aecaee1805ef4109aadd433e7
+
 FREETYPE_HREF="http://download.savannah.gnu.org/releases/freetype/freetype-2.5.3.tar.gz"
+FREETYPE_MD5=cafe9f210e45360279c730d27bf071e9
+
 FONTCONFIG_HREF="http://www.freedesktop.org/software/fontconfig/release/fontconfig-2.11.1.tar.gz"
+FONTCONFIG_MD5=e75e303b4f7756c2b16203a57ac87eba
 
 cd $WORKDIR
 
@@ -31,8 +42,12 @@ export PKG_CONFIG_PATH=$INSTDIR/lib/pkgconfig:/usr/lib/pkgconfig
 
 function download {
     cd $DLDIR
-    echo "downloading $1"
-    curl -L -C - -o $1.tar.gz "$2"
+
+    if [ ! -e "$1" ] || [ `md5sum $1.tar.gz` -ne "$3" ] then
+	rm -f $1.tar.gz
+    	echo "downloading $1"
+	curl -L -C - -o $1.tar.gz "$2"
+    fi
 }
 
 function build {
@@ -60,12 +75,12 @@ echo "======================================================================="
 echo "= downloading libraries ..."
 echo "======================================================================="
 
-download graphviz   "$GRAPHVIZ_HREF"
-download libgd      "$LIBGD_HREF"
-download libpng     "$LIBPNG_HREF"
-download libjpeg    "$LIBJPEG_HREF"
-download fontconfig "$FONTCONFIG_HREF"
-download freetype   "$FREETYPE_HREF"
+download graphviz   "$GRAPHVIZ_HREF"   "$GRAPHVIZ_MD5"
+download libgd      "$LIBGD_HREF"      "$LIBGD_MD5"
+download libpng     "$LIBPNG_HREF"     "$LIBPNG_MD5"
+download libjpeg    "$LIBJPEG_HREF"    "$LIBJPEG_MD5"
+download fontconfig "$FONTCONFIG_HREF" "$FONTCONFIG_MD5"
+download freetype   "$FREETYPE_HREF"   "$FREETYPE_MD5"
 
 echo "======================================================================="
 echo "= building libraries ..."
